+++ backend/src/models/User.model.js
@@ -59,6 +59,27 @@ const userSchema = new mongoose.Schema(
   },
 );
 
-const User = mongoose.model("User", userSchema);
+const archivedUserSchema = new mongoose.Schema(
+  {
+    original_user_id: {
+      type: mongoose.Schema.Types.ObjectId,
+      required: true,
+    },
+    name: { type: String },
+    email: { type: String },
+    role: { type: String },
+    deleted_at: {
+      type: Date,
+      default: Date.now,
+      expires: 2592000, //MongoDB automatically deletes this document 30Days after deleted_at
+    },
+  },
+  {
+    timestamps: { createdAt: "created_at", updatedAt: "updated_at" },
+  },
+);
 
-module.exports = User;
+const User = mongoose.model("User", userSchema);
+const ArchivedUser = mongoose.model("ArchivedUser", archivedUserSchema);
+
+module.exports = { User, ArchivedUser };